Converts documentation to markdown
Instead of using Sphinx, documentation is now written in Markdown (specifically,
github-flavored markdown), and can be rendered using either MkDocs or Bookdown.

The Pragmas section was updated to pull in all shipped pragma documentation from
the README as well.

This commit also removes support for serving the API documentation via RTD; I'm
not sure it's useful, particularly as it easily goes out-of-date. Users can
build it on their own if desired.